Use PIL for image formats
This fixes a bug where "webm" format images, which download with the extension ".jpg" cause an error during training. Now check_images.py uses PIL to get the image type, and images that are not JPEG format are removed from the image directory.
Also update .gitignore for the downloads folder.